Implementação do TopBar para navegar para lista de postas
