+++ app/src/main/java/com/marcelo/loteriadossonhos/ui/theme/Color.kt
@@ -13,3 +13,4 @@ val Pink40 = Color(0xFF7D5260)
 val Green = Color(0xFF00AA69)
 val White = Color(0xFFFFFFFF)
 val Blue = Color(0xFF260085)
+val BlueLigth = Color(0xFF455CAE)
